DRTVWR-476: Allow more control over debugging output destination.
Instead of testing #if LOG_DEVICES, LOG_ELEMENTS or NDOF_DEBUG, in other words
using them as booleans, make each of those (plus LOG_SCORING and
LOG_SEARCHING, which appear to remain unused) designate a FILE* stream of
interest (stdout or stderr). Test for whether they're defined at all. Use the
value of the applicable symbol as the output stream for the debugging output
controlled by (the presence of) that symbol. Pass target FILE* into a few
functions that used to produce output unconditionally to either stdout or
stderr.

Since a couple of .vcxproj files set NDOF_DEBUG, make them set it to 'stderr'
instead of to 1.